dish-ran-dish-dev/requirements.txt
#NOTE: python3.11 needs to be used
uvicorn==0.34.0
fastapi==0.115.6
pytz==2024.2
langgraph==0.2.60
langchain==0.3.14
langchain_ibm==0.3.5
python-dotenv==1.0.1
ibm_cloud_sdk_core==3.22.0
ibm_watson==9.0.0
boto3==1.37.3
pyarrow==18.1.0
psycopg==3.2.3
psycopg_pool==3.2.4
rich==13.9.4
psycopg-binary==3.2.3
langgraph-checkpoint==2.0.9
langgraph-checkpoint-postgres==2.0.9
langgraph-sdk==0.1.48
psycopg2-binary==2.9.10
langchain_community==0.3.14
pypdf==5.1.0
# snowflake-connector-python==3.12.4
snowflake-connector-python==3.15.0
snowflake-sqlalchemy==1.7.5
SQLAlchemy==2.0.41
sqlalchemy-orm==1.2.10
async-timeout==4.0.3
asyncpg==0.29.0
httpx==0.27.2
numpy==1.26.4
pandas==2.1.4
aiofiles==24.1.0
openpyxl==3.1.5
ibm-watson-openscale==3.0.2
ibm-cos-sdk-core==2.13.6
redis[async]
llama-index-core
llama-index
llama-index-llms-ibm
python-dateutil
rapidfuzz
aioboto3==14.3.0
langchain-mcp-adapters==0.1.1
scikit-learn==1.7

markdown-pdf==1.9
PyMuPDF==1.26.4
markdown==3.9